requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.7
lxml==4.9.3 